Create shiny module for choosing an audio directory and selecting which files to import

Create shiny module for choosing an audio directory and selecting which files to import

import_audio_ui(
  id,
  title = "Select a folder containing audio files",
  box_title = "Set Audio Library"
)

import_audio_server(id, audio_dir = NULL, title = "Starting Directory")

Arguments

id

module id

title

label fed to shinyDirButton

box_title

label of the first shinydashboardPlus::box

audio_dir

an initial audio directory

status

the color of the box. See shinydashboardPlus::box status

collapsible

whether the module can be collapsed/hidden

collapsed

whether the module should start off as collapsed

Functions

import_audio_ui(): Create module UI for choosing an audio directory and selecting which files to import

import_audio_server(): Create module server for choosing an audio directory and selecting which files to import