Make howler UI for use in audio_playpack_ui

Make howler UI for use in audio_playpack_ui

make_howler_ui(
  files,
  howler_id = "howler",
  seek_ping_rate = 100,
  include_current_track = TRUE,
  width = NULL,
  options = list(),
  ...
)

Arguments

files

Files that will be used in the player.

howler_id

ID to correspond to the howler object and its related buttons.

seek_ping_rate

Number of milliseconds between each update of input${id}_seek while playing. Default is set to 1000. If set to 0, then input${id}_seek will not exist. Decreasing this number too much could increase the cross-talk between the server and client to the point where significant lags occur.

include_current_track

Logical, should the current track be included in the UI of the module?

width

Width (in pixels) of the player. Defaults to 300px.