Append a symbol to a dataframe. This is the title.

Adds the chosen symbol as a column to the end of a dataframe, and optionally create an ID column. This will be the description.

my_function(data, symbol = c("a", "b"), add_id_col = FALSE, group_by_col)

Arguments

data

a dataset. Defaults to mtcars.

symbol

the symbol to append

add_id_col

Logical (TRUE/FALSE). If TRUE, add an ID column and append it to the front of the dataframe.

group_by_col

column to group by if creating an ID column.

Value

a dataframe

Details

This is a terrible, disgusting function. Why would you ever want to do this. Im mainly trying to illustrate the different types of arguments/defaults you can have, and how roxygen/the loaded function will work.

The @importFrom statements above will not show up in the helper file (this text will), however the NAMESPACE file will update when you run devtools::document() if you modify them.

The @importFrom statements above translate to:

"import syms and sym from the package rlang

"import n from the package dplyr

bolded word

italizized phrase

this phrase will appear as code

This section will be a code block:

# Adds a new column to a dataset, repeating 'A' all the way down (see ?dplyr::mutate for more details):
data %>% dplyr::mutate(new_column = "A")

Note

mtcars is an example dataset exported by the datasets package, and comes with the base installation of R.

Examples

if (FALSE) {
# Defaults to symbol 'a'
my_function(data = mtcars)

# Defaults to `mtcars` dataset
my_function(data = mtcars, symbol = "b")

# Create ID and row_number columns
my_function(data = mtcars, add_id_col = TRUE, group_by_col = "gear")

# Would error out in a controlled way:
my_function(data = mtcars, add_id_col = TRUE)
my_function(data = mtcars, add_id_col = TRUE, group_by_col = "columnDoesntExist")
my_function(data = mtcars, symbol = "c")
}