Plot wave channels of audio file

Plot wave channels of audio file

New plotting method

Original plotting method

Process Wave channel

plot_wave_audio(
  audio_obj,
  source = "wave_audio",
  type = c("left", "right", "stereo"),
  format = c("fancy", "base"),
  include_info = FALSE,
  range_slider = TRUE,
  inner_plot = TRUE,
  hollow = FALSE,
  simplify = TRUE,
  nr = 4000,
  plot_title = NULL,
  xunit = c("Time", "Samples"),
  ylim = NULL,
  xlab = NULL,
  ylab = NULL
)

plot_wave_channel_fancy(
  audio_data,
  audio_params,
  source = NULL,
  ylim = NULL,
  xlab = "Time",
  ylab = "Audio Channel",
  hollow = FALSE,
  show_y_axis = FALSE,
  line_color = "#1b5fa6",
  ft_color = "lightgrey",
  bg_color = "#252525",
  range_slider = TRUE,
  inner_plot = TRUE,
  inner_line_color1 = "#4d91c8",
  inner_line_color2 = "#75b8ed"
)

plot_wave_channel_base(
  audio_data,
  audio_params,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  plot_title = NULL,
  axes = TRUE,
  center = TRUE
)

process_wave_channel(audio_obj, xunit = "Time", simplify = TRUE, nr = 2500)

Arguments

audio_obj

a tuneR audio object.

source

a character string of length 1. Match the value of this string with the source argument in plotly::plot_ly() to respond to events emitted from that specific plot. Only used if format = "fancy". See ?plotly::event_data for more details.

type

channel to plot. One of c("left", "right", "stereo").

format

format type. "fancy" is experimental and will eventually be the main one used within the app.

include_info

Logical (TRUE/FALSE). If TRUE, append information about the audio file.

range_slider

Logical (TRUE/FALSE). If TRUE, add a rangeslider.

inner_plot

Logical (TRUE/FALSE). If TRUE, add an inner plot consisting of two lines.

hollow

Logical (TRUE/FALSE). If TRUE, make the plot hollow. Only usable for "fancy" plots.

simplify

Logical (TRUE/FALSE). Whether to simplify large audio files. Defaults to TRUE.

nr

Number of lines to draw for simplified plots. Only takes affect if simplify = TRUE.

plot_title

a title for the plot (optional).

xunit

unit for x-axis. One of c("Time", "Samples").

ylim

y-limit. Will be determined if not set.

xlab

label for x-axis. Defaults to xunit if not set.

ylab

label for y-axis. Will be determined if not set.

audio_data

Audio data returned from process_wave_channel.

audio_params

List of audio parameters returned from process_wave_channel.

show_y_axis

Logical (TRUE/FALSE). If TRUE, show the y-axis (limits only).

line_color

line color

ft_color

font color

bg_color

background color

inner_line_color1

line color of inner line 1

inner_line_color2

line color of inner line 2 (smallest of the three)

axes

Logical (TRUE/FALSE). If TRUE, add axes to the plot

center

Logical (TRUE/FALSE). If TRUE, center the plot

Value

a plotly object if format = "fancy". Otherwise NULL invisibly

Details

plot_wave_channel_base and process_wave_channel were inspired by tuneR's plotting method for wave files: https://github.com/cran/tuneR/blob/master/R/plot-Wave.R

We want more control over the visual, the ability to plot the different wave channels, and for the plot to be interactive, so this code had to be pulled out and refactored. plot_wave_channel_fancy uses plotly to create a much nicer plot with additional features.

Not sure if source will be supported for stereo types

See also

Examples

if (FALSE) {
# Read in audio file with `tuneR`
audio_obj <- tuneR::readMP3(file.path(EXAMPLE_AUDIO_DIR, "flowers.mp3"))

# Plot variations
plot_wave_audio(audio_obj)
plot_wave_audio(audio_obj, format = "base") # similar to `tuneR::plot()`
plot_wave_audio(audio_obj, type = "right")
plot_wave_audio(audio_obj, type = "stereo")
plot_wave_audio(audio_obj, range_slider = FALSE, include_info = FALSE)

# step-wise (can only process one channel at a time - defaults to `audio_obj@left`)
wave_channel <- process_wave_channel(audio_obj, simplify = TRUE, nr = 2500)
plot_wave_channel_fancy(
    audio_data = wave_channel$audio_data,
    audio_params = wave_channel$params
)
}